[PP] amount should be typed as BigDecimal (#26)

+++ cor/src/main/scala/timetopaytaxpayer/cor/model/Debit.scala
@@ -22,7 +22,7 @@ import play.api.libs.json.{Json, OFormat}
 
 final case class Debit(
     originCode: String,
-    amount:     Double,
+    amount:     BigDecimal,
     dueDate:    LocalDate,
     interest:   Option[Interest],
     taxYearEnd: LocalDate
